requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
astroid==3.3.6
attrs==24.3.0
autopep8==2.3.1
awscli==1.36.36
awsebcli==3.21.0
black==24.10.0
blessed==1.20.0
boto3==1.35.95
botocore==1.35.95
cachetools==6.0.0
cement==2.10.14
certifi==2024.8.30
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.9
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
django-storages==1.14.4
django-stubs-ext==5.2.0
django-tasks==0.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.16
drf-yasg==1.21.8
factory_boy==3.3.1
Faker==33.1.0
filelock==3.16.1
flake8==7.1.1
foreman==0.9.7
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
identify==2.6.3
idna==3.10
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.2
pathspec==0.10.1
platformdirs==4.3.6
pre_commit==4.0.1
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.0
pylint==3.3.2
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
qrcode==8.2
redis==5.2.1
requests==2.32.3
requests-futures==1.0.2
rsa==4.7.2
s3transfer==0.10.4
semantic-version==2.10.0
sendgrid==6.11.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
termcolor==2.5.0
tomlkit==0.13.2
twilio==9.4.1
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
virtualenv==20.28.0
wcwidth==0.2.13
websocket==0.2.1
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
